Updated the DispatchQueue extension to follow the suggestion from https://github.com/yannickl/AwaitKit/issues/9. Updated the version to 2.1.0.
