+++ Sources/AwaitKit.swift
@@ -38,11 +38,31 @@ public struct Queue {
     static let await = DispatchQueue(label: "com.yannickloriot.awaitqueue", attributes: .concurrent)
 }
 
-// MARK: - Async -
+public final class Extension<Base> {
+    public let base: Base
+    public init(_ base: Base) {
+        self.base = base
+    }
+}
+
+/**
+ A type that has AwaitKit extensions.
+ */
+public protocol AwaitKitCompatible {
+    associatedtype CompatibleType
+    var ak: CompatibleType { get }
+}
+
+public extension AwaitKitCompatible {
+    public var ak: Extension<Self> {
+        get { return Extension(self) }
+    }
+}
+
+extension DispatchQueue: AwaitKitCompatible { }
 
 /**
  Yields the execution to the given closure and returns a new promise.
- 
  - parameter body: The closure that is executed on a concurrent queue.
  - returns: A new promise that is resolved when the provided closure returned.
  */
@@ -52,113 +72,30 @@ public func async<T>(_ body: @escaping () throws -> T) -> Promise<T> {
 
 /**
  Yields the execution to the given closure which returns nothing.
- 
  - parameter body: The closure that is executed on a concurrent queue.
  */
 public func async(_ body: @escaping () throws -> Void) {
-    async(body, queue: Queue.async)
+    Queue.async.ak.async(body)
 }
-
-/**
- Yields the execution to the given closure and returns a new promise.
- 
- - Parameter body: The closure that is executed on the given queue.
- - Returns: A new promise that is resolved when the provided closure returned.
- */
-public func async<T>(_ body: @escaping () throws -> T, queue: DispatchQueue) -> Promise<T> {
-    return queue.promise(execute: body)
-}
-
-/**
- Yields the execution to the given closure which returns nothing.
- 
- - Parameter body: The closure that is executed on the given queue.
- */
-public func async(_ body: @escaping () throws -> Void, queue: DispatchQueue) {
-    let promise: Promise<Void> = async(body)
-    
-    promise.catch { _ in }
-}
-
-// MARK: - Await -
 
 /**
  Awaits that the given closure finished and returns its value or throws an error if the closure failed.
- 
  - parameter body: The closure that is executed on a concurrent queue.
  - throws: The error sent by the closure.
  - returns: The value of the closure when it is done.
  */
 @discardableResult
 public func await<T>(_ body: @escaping () throws -> T) throws -> T {
-    return try await(body, queue: Queue.await)
+    return try Queue.await.ak.await(body)
 }
 
 /**
  Awaits that the given promise resolved and returns its value or throws an error if the promise failed.
- 
  - parameter promise: The promise to resolve.
  - throws: The error produced when the promise is rejected.
  - returns: The value of the promise when it is resolved.
  */
 @discardableResult
 public func await<T>(_ promise: Promise<T>) throws -> T {
-    return try await(promise, queue: Queue.await)
-}
-
-/**
- Awaits that the given closure finished on the receiver and returns its value or throws an error if the closure failed.
- 
- - parameter body: The closure that is executed on the receiver.
- - throws: The error sent by the closure.
- - returns: The value of the closure when it is done.
- - seeAlso: await(promise:)
- */
-@discardableResult
-public func await<T>(_ body: @escaping () throws -> T, queue: DispatchQueue) throws -> T {
-    let promise = queue.promise(execute: body)
-    
-    return try await(promise, queue: queue)
-}
-
-/**
- Awaits that the given promise resolved on the receiver and returns its value or throws an error if the promise failed.
- 
- - parameter promise: The promise to resolve.
- - throws: The error produced when the promise is rejected or when the queues are the same.
- - returns: The value of the promise when it is resolved.
- */
-@discardableResult
-public func await<T>(_ promise: Promise<T>, queue: DispatchQueue) throws -> T {
-    guard queue.label != DispatchQueue.main.label else {
-        throw NSError(domain: "com.yannickloriot.awaitkit", code: 0, userInfo: [
-            NSLocalizedDescriptionKey: "Operation was aborted.",
-            NSLocalizedFailureReasonErrorKey: "The current and target queues are the same."
-            ])
-    }
-    
-    var result: T?
-    var error: Swift.Error?
-    
-    let semaphore = DispatchSemaphore(value: 0)
-    
-    promise
-        .then(on: queue) { value -> Void in
-            result = value
-            
-            semaphore.signal()
-        }
-        .catch(on: queue) { err in
-            error = err
-            
-            semaphore.signal()
-    }
-    
-    _ = semaphore.wait(timeout: DispatchTime(uptimeNanoseconds: UINT64_MAX))
-    
-    guard let unwrappedResult = result else {
-        throw error ?? Error.general
-    }
-    
-    return unwrappedResult
+    return try Queue.await.ak.await(promise)
 }
